Get julia multithreaded, be a little clever
